Fixed incorrect types for svcBreak
svcBreak reason should be a u32, not a u64.

+++ src/core/hle/kernel/svc.cpp
@@ -303,13 +303,13 @@ static ResultCode ArbitrateUnlock(VAddr mutex_addr) {
 
 struct BreakReason {
     union {
-        u64 raw;
-        BitField<31, 1, u64> dont_kill_application;
+        u32 raw;
+        BitField<31, 1, u32> dont_kill_application;
     };
 };
 
 /// Break program execution
-static void Break(u64 reason, u64 info1, u64 info2) {
+static void Break(u32 reason, u64 info1, u64 info2) {
     BreakReason break_reason{reason};
     if (break_reason.dont_kill_application) {
         LOG_ERROR(
